fix(rtc): CRITICAL - Cast drift_buffer_count to int64_t in division to preserve sign
Problem: Dividing int64_t by size_t (unsigned) causes implicit conversion
When error_change_avg_ns (int64_t) is divided by drift_buffer_count (size_t),
the signed value gets promoted to unsigned, causing wraparound for negative results.

Example:
  Buffer: [-2556, +852]
  Sum: -1704
  Average should be: -1704 / 2 = -852 ns
  But displayed: 9223372036854774956 ns (wraparound!)

Root cause: Line 689 had error_change_avg_ns /= drift_buffer_count
This performed SIGNED / UNSIGNED division  implicit unsigned conversion

Solution: Cast to int64_t before division:
  error_change_avg_ns /= static_cast<int64_t>(drift_buffer_count);

This was the REAL bug Expert predicted - mixed signed/unsigned arithmetic!
Previous commit 73225d4 fixed the dual-calculation issue but missed THIS.

Expert Analysis (deb.md): 100% correct about signed/unsigned conversion
Evidence (console.log):
  Avg(2): -0.085 ppm (9223372036854774956ns/10s)  wraparound
  Should be: -0.085 ppm (-852ns/10s)

Thank you Expert for the precise diagnosis!

+++ examples/12-RasPi5_i226-grandmaster/src/ptp_grandmaster.cpp
@@ -686,7 +686,7 @@ int main(int argc, char* argv[])
                             for (size_t i = 0; i < drift_buffer_count; i++) {
                                 error_change_avg_ns += error_change_buffer[i];
                             }
-                            error_change_avg_ns /= drift_buffer_count;
+                            error_change_avg_ns /= static_cast<int64_t>(drift_buffer_count);  // Cast to preserve sign
                             current_drift_ppm = drift_ppm;
                             current_drift_avg = drift_avg;
                             current_error_change_avg_ns = error_change_avg_ns;  // Store for reuse
